refactoring to take in account some remarks from esgn
