+++ .env.example
@@ -15,5 +15,5 @@ POSTGRES_SCHEMA=edition211
 
 # Valeurs de configuration de l'API
 API_PORT=8010
-LIMIT_FEATURE=1
+# ALL or number
 MAX_FEATURE=1000
