--- DOFIRST.md
@@ -1,24 +1,0 @@
-# More actions
-
-> 🚨 This is for python 3.9. Please adapt for your own version.
-> Use this only if you need to make some local tests without docker.
-
-## Create a python env
-
-> See https://docs.python.org/3/tutorial/venv.html
-
-```
-python3 -m venv ign-env
-source ign-env/bin/activate
-```
-
-if first time :
-```
-pip install --upgrade pip
-pip install -r requirements.txt
-```
-
-## TODO ?
-- [ ] Add a progress bar in downloader python script
-- [ ] Add the possibility to select the version of the parcellaire express
-- [ ] Add welcome data
