+++ api/app.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 	"log"
 	"net/http"
+	"os"
 
 	"github.com/gorilla/mux"
 	_ "github.com/lib/pq"
@@ -31,9 +32,7 @@ type App struct {
 	DB     *sql.DB
 }
 
-func (a *App) Initialize() {
-
-	// log.SetOutput(os.Stderr)
+func (a *App) Initialize(DB *sql.DB) {
 
 	a.DB = DB
 
@@ -117,7 +116,11 @@ func (a App) healthCheckHandler(w http.ResponseWriter, r *http.Request) {
 }
 
 func (a *App) initializeRoutes() {
-	a.Router.PathPrefix("/viewer/").Handler(http.StripPrefix("/viewer/", http.FileServer(http.Dir("./views")))).Methods("GET")
+	theViewerUrl, isViewerUrldefined := os.LookupEnv("VIEWER_URL")
+	if isViewerUrldefined {
+		log.Printf("Html viewer is enabled : %v", isViewerUrldefined)
+		a.Router.PathPrefix(theViewerUrl).Handler(http.StripPrefix(theViewerUrl, http.FileServer(http.Dir("./views")))).Methods("GET")
+	}
 
 	a.Router.Handle("/parcelle/{idu:"+iduRegex+"}", Use(LogMw).ThenFunc(a.getById)).Methods("GET")
